Reduced the need for openlayers to be compiled EVERY time you start,
Now it will only compile it if it does not exist in the build dir.

This will speed up a HUGE amount of the build process.

+++ start.sh
@@ -9,10 +9,12 @@ node_modules/.bin/lessc src/less/signin.less public/css/signin.css &&
 node_modules/.bin/lessc src/less/charts.less public/css/charts.css &&
 node_modules/.bin/lessc src/less/maptest.less public/css/maptest.css &&
 
-cd public/lib/openlayers/build &&
-python build.py &&
-cp -r OpenLayers.js ../../../../public/lib/dependencies/js/ &&
-cd ../../../../ &&
+if [ ! -f public/lib/openlayers/build/OpenLayers.js ]; then
+	cd public/lib/openlayers/build &&
+	python build.py &&
+	cp -r OpenLayers.js ../../../../public/lib/dependencies/js/ &&
+	cd ../../../../
+fi
 
 echo 'BUILDING GOGRAM' &&
 go build &&
